requirements.txt
websockets==10.4
cryptography==3.4.7
PyJWT==2.3.0
bcrypt==3.2.0